+++ src/tubeviz/__init__.py
@@ -1,2 +1,2 @@
 # SPDX-License-Identifier: Apache-2.0
-__version__ = "0.33.6"
+__version__ = "0.33.7"